requirements.txt
streamlit
google-generativeai
google-cloud-storage
langchain
langchain-google-genai
python-dotenv 